requirements.txt
click
musicbrainzngs
mutagen
thefuzz
python-Levenshtein
tqdm
inquirer
colorama
requests
spotipy
python-dotenv 